+++ lib/semantic_navigation/core/render.rb
@@ -3,6 +3,7 @@ module SemanticNavigation
     module Render
       
       def render
+        return nil if !render_item?
         if @parent
           sub = @active || show_submenu? ? render_submenu : nil
           link = view_object.link_to @name, @url_options, :class => classes, :id => a_id
@@ -19,7 +20,7 @@ module SemanticNavigation
       
      def active_item_name
        if @active && @sub_items.count > 0
-        name = @sub_items.map{|s| s.active_item_name}.find{|s| s != nil}
+         name = @sub_items.map{|s| s.active_item_name}.find{|s| s != nil}
        end
        if active?
          name = @name  
@@ -32,14 +33,13 @@ module SemanticNavigation
       def render_submenu
         if @sub_items.count > 0
           sub = @sub_items.map{|s| s.render}.sum
-          view_object.content_tag(:ul, sub, :id => ul_id)
+          view_object.content_tag(:ul, sub, :id => ul_id) if sub != 0
         end
       end
     
       def active?
         view_object.current_page?(@url_options) if !@url_options.nil?
       end
-      
     end  
   end
 end