+++ Changelog.md
@@ -3,6 +3,9 @@
 * Possibility to define default render options (active_class and show_ids)
 * Possibility to define the private render options to each menu item
 * Possibility to render or not to render submenu if parent item is inactive
+* Catch of undefined helper methods
+* Catch of undefined menu render calls
+* Render active name submenu
 
 ### 0.0.1 / December 23th, 2011