+++ Changelog.md
@@ -10,6 +10,7 @@
 * Changed helper method to render menu
 * Added prefixes for menu, item, name ids
 * Added conditions for rendering (function show_if)
+* Added breadcrumbs rendering
 
 ### 0.0.1 / December 23th, 2011